requirements.txt
click==7.1.1
Jinja2==2.11.1
MarkupSafe==1.1.1
pathtools==0.1.2
PyYAML==5.3.1
watchdog==0.10.2
